refactor(notifications): extract notification types to types.rs
Move SynchronizationState, Warning, ValidationMode, and
BlockValidationResult enums to dedicated types.rs module.
