+++ src/notifications/mod.rs
@@ -1,6 +1,9 @@
 pub mod notification;
+pub mod types;
 
 pub use notification::{
     BlockTip, FatalError, FlushError, HeaderTip, KernelNotificationInterfaceCallbacks, Progress,
     WarningSet, WarningUnset,
 };
+
+pub use types::{BlockValidationResult, SynchronizationState, ValidationMode, Warning};
